Invert boolean display logic for overdue skill trees on admin users index
This simply flips the boolean logic / wording for whether team members
have a current skill tree. This allows ActiveAdmin to render the yes/no
labels in the appropriate color for the situation (right now it's a little
visually confusing because you are shown green labels for the "bad" status).

With this change:

green = good (the user has a current skill tree)
yellow = uh oh (the user's skill tree is out of date)