Configure Heroku CI to run tests on branch update
This enables automatic running of the test suite for Stacks via Heroku CI
whenever a branch ref is updated in Github.

In order for this to work I needed to add an app.json manifest to tell
Heroku that A) we want to use a Ruby buildpack, and B) to configure an
ephemeral Postgres dyno that we can connect to from the test runner.

I also had to adjust a few fixtures to address some failing non-null constraints.