Always show associated forecast projects for project tracker, even if associated with another
This fixes a display issue in the UI for editing an existing project tracker.
Currently, if the same forecast project has been associated with multiple
project trackers, its name will not be displayed in the dropdown for the
forecast project section in any of the associated trackers (even though
it's actually associated). This creates unnecessary confusion.

This should only affect historical projects, since validation was added
to prevent multiple trackers from being associated with the same forecast project
going forward.